Initial commit
Create a very simple project that uses Valhalla, and contains a simple
tool that generates a list of OSM way_id's from start and end points
(and a Valhalla configuration). This is a proof-of-concept as much as
anything, although there is a fair chance the code will grow into
something more directly useful.