Add 'convert' "tool"
Add a "tool", which is really more of a proof-of-concept, to read OSM
and convert it into a simplified format which we will hopefully be able
to use to convert Valhalla output into node lists. This is necessary
because just getting the way ids is not sufficient for our ultimate
objective, which is generating routes for SUMO.

Based on timing tests, there does not seem to be significant value in
performing this process offline, so likely this logic will be folded
into a later tool to generate routes (with the assumption that such tool
will generate many routes in one run).