Implement way segmentation
Add a method in Graph to split a way into segments at intersections, and
optionally to also split loops. Use this in the 'trace' tool to try to
guess how to map a route into SUMO edges (essentially, we are trying to
replicate how SUMO splits ways in order to guess how it will assign
edges).

This should plausibly get us to the point that we can spit out a set of
SUMO edge identifiers, provided that the SUMO network has not been
modified since conversion from OSM.