Clean up build logic
Remove a bunch of unused or unneeded stuff from the build logic and
refactor some other stuff into separate files so that the root
CMakeLists.txt is much simpler. Fix a number of issues (wrong install
locations, wrong version and namespace due to copying from a different
project).

This is currently broken because neither readosm or tbtrace have been
fixed. These both have their own issues that need to be addressed as
well; readosm shouldn't be bundled at all, and tbtrace doesn't belong in
the directory for the tbutil library.