"Clean up" includes
Change interface include directories to more closely match what we
expect an installation to look like. This also forces us, at least to
the extent our headers are used outside of their immediate target, to
prefix our includes the same way an external user would, which is a good
thing.