Move node lookup to routing
Move the code to look up OSM nodes from the Valhalla trip into the
function that computes the route. This is a precursor step to moving
said function to the tbutil library, as returning a set of nodes is more
sensible from an API standpoint than returning the raw locations.
However, for the sake of avoiding any potential data loss, the raw
locations are still included.