Add search capability to RoutingEngine
Add a new method to the recently refactored RoutingEngine to allow
looking up a point in the routing data. This provides a way to correlate
an arbitrary input to a defined location, which is particularly useful
for trying to determine if a location is anywhere near the available
data set.