Rewrite CSV reading code
Rewrite the code to read CSV files as a stream reader. For large files,
we really don't want to read the entire file contents into memory in
loosely structured form, when we are almost certainly going to convert
to a more structured form. (Ideally, we would not ever load the entire
file into memory even in final, fully structured form, but that may not
be possible. This at least lets us cut out a completely unneeded
intermediate representation.)

Also, I have no idea what structure the old logic was expecting. (Pair
of string and vector-of-float? Huh?) The new logic makes no assumptions
about the structure of the file while still providing convenience
methods to perform on-the-fly data conversions. (Right now, only string
and double are supported, but more types can be added easily as needed.)