Add function to convert timestamp string
Create a utility function to convert an ISO date/time string into a
standard format (in this case, seconds since the UNIX epoch). This is
essentially equivalent to the non-portable strptime. (Even converting
from a time struct in UTC is not portable; ugh! Accordingly, there is
necessarily a small amount of platform-specific code in this function.)