Set default <project>_RUNTIME_PREFIX in superbuild
When building as a superbuild, set <project>_RUNTIME_PREFIX by default
to the superbuild's install prefix. This makes sense since just building
the superbuild installs to here, and the superbuild install prefix is
usually (indeed, defaults to) a subdirectory of the build directory
rather than an "actual" install. Particularly, since in this
configuration it is likely expected that everything is run out of the
superbuild install tree, this makes it possible to build individual
targets (at least for our own project) without having to re-run a full
install of ourselves.