Fix waypoint culling in tbroute
Modify the algorithm in tbroute that culls "too close together"
waypoints to not remove the end waypoints of a leg. This avoids a
degenerate case that could cause multiple points at the start of a leg
to be incorrectly thrown out.

The trouble is that the algorithm identifies the shortest segment in a
leg, but then needs to decide somehow which point to remove. We never
want to remove the end of a leg, but for other cases, we need to decide
somehow. The new algorithm considers the distance between the two
adjacent points for each of the points of the shortest segment and
removes the point between the shorter of said pairs.